+++ dart/jovial_misc/lib/io_utils.dart
@@ -21,10 +21,10 @@ class EOFException implements IOException {
   }
 }
 
-/// DataInputStream is a coroutines-style wrapper around a
-/// Stream<List<int>>, like you get from a socket or a file in Dart.
+/// `DataInputStream` is a coroutines-style wrapper around a
+/// `Stream<List<int>>`, like you get from a socket or a file in Dart.
 /// This lets you asynchronously read a stream using an API much like
-/// java.io.DataInputStream.
+/// `java.io.DataInputStream`.
 ///
 /// Note that this class does not read all of the data into a buffer first,
 /// so it is very space-efficient.  It does, however, use asynchrony at a
@@ -32,7 +32,7 @@ class EOFException implements IOException {
 /// on how efficient async/await are in current Dart implementations.  As
 /// of this writing (Jan. 2020), I have not measured.
 ///
-/// See also ByteBufferDataInputStream.
+/// See also [ByteBufferDataInputStream].
 class DataInputStream {
   // This class iterates through the provided Stream<List<int>> directly,
   // and buffers by hanging onto a Uint8List  until it's been completely
@@ -42,11 +42,13 @@ class DataInputStream {
   final StreamIterator<List<int>> _source;
   Uint8List _curr;
   int _pos;
-  static const utf8Decoder = Utf8Decoder(allowMalformed: true);
+  static const _utf8Decoder = Utf8Decoder(allowMalformed: true);
 
-  DataInputStream._internal(this._source);
+  /// The current endian setting, either [Endian.big] or [Endian.little].
+  /// This is used for converting numeric types.
+  Endian endian;
 
-  DataInputStream(Stream<List<int>> source)
+  DataInputStream(Stream<List<int>> source, [this.endian = Endian.big])
       : _source = StreamIterator<List<int>>(source);
 
   /// Returns the number of bytes that can be read from the internal buffer.
@@ -88,10 +90,10 @@ class DataInputStream {
     // isEOF() sets up _curr and _pos as a side effect.
   }
 
-  /// Returns a new, mutable Uint8List containing the desired number
+  /// Returns a new, mutable [Uint8List] containing the desired number
   /// of bytes.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<Uint8List> readBytes(int num) async {
     if (num == 0) {
       return Uint8List(0);
@@ -113,10 +115,10 @@ class DataInputStream {
     }
   }
 
-  /// Returns a potentially immutable Uint8List containing the desired number
+  /// Returns a potentially immutable [Uint8List] containing the desired number
   /// of bytes.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<Uint8List> readBytesImmutable(int num) async {
     if (num == 0) {
       return Uint8List(0);
@@ -141,65 +143,69 @@ class DataInputStream {
     }
   }
 
-  /// Reads and returns a 4-byte signed big-endian integer.
+  /// Reads and returns a 4-byte signed integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readInt() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 4) {
+    if (_curr.length - _pos < 4) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(4)).buffer).getInt32(0);
+      return ByteData.view((await readBytes(4)).buffer).getInt32(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getInt32(_pos);
+      var result = ByteData.view(_curr.buffer).getInt32(_pos, endian);
       _pos += 4;
       return result;
     }
   }
 
-  /// Reads and returns a 4-byte unsigned big-endian integer.
+  /// Reads and returns a 4-byte unsigned integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readUnsignedInt() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 4) {
+    if (_curr.length - _pos < 4) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(4)).buffer).getUint32(0);
+      return ByteData.view((await readBytes(4)).buffer).getUint32(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getUint32(_pos);
+      var result = ByteData.view(_curr.buffer).getUint32(_pos, endian);
       _pos += 4;
       return result;
     }
   }
 
-  /// Reads and returns a 2-byte unsigned big-endian integer.
+  /// Reads and returns a 2-byte unsigned integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readUnsignedShort() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 2) {
+    if (_curr.length - _pos < 2) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(2)).buffer).getUint16(0);
+      return ByteData.view((await readBytes(2)).buffer).getUint16(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getUint16(_pos);
+      var result = ByteData.view(_curr.buffer).getUint16(_pos, endian);
       _pos += 2;
       return result;
     }
   }
 
-  /// Reads and returns a 2-byte signed big-endian integer.
+  /// Reads and returns a 2-byte signed integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readShort() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 2) {
+    if (_curr.length - _pos < 2) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(2)).buffer).getInt16(0);
+      return ByteData.view((await readBytes(2)).buffer).getInt16(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getInt16(_pos);
+      var result = ByteData.view(_curr.buffer).getInt16(_pos, endian);
       _pos += 2;
       return result;
     }
@@ -207,13 +213,13 @@ class DataInputStream {
 
   /// Reads and returns an unsigned byte.
   ///
-  /// Throws EOFException if EOF is reached before the needed byte is read.
+  /// Throws [EOFException] if EOF is reached before the needed byte is read.
   Future<int> readUnisgnedByte() async {
     await _ensureNext();
     return _curr[_pos++];
   }
 
-  /// Reads and returns a signed byte.
+  /// Reads and returns a signed byte.  No sign extension is done.
   ///
   /// Throws EOFException if EOF is reached before the needed byte is read.
   Future<int> readByte() async {
@@ -226,54 +232,90 @@ class DataInputStream {
     }
   }
 
-  /// Reads and returns an 8-byte signed big-endian integer.
+  /// Reads and returns an 8-byte integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readLong() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 8) {
+    if (_curr.length - _pos < 8) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(8)).buffer).getInt64(0);
+      return ByteData.view((await readBytes(8)).buffer).getInt64(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getInt64(_pos);
+      var result = ByteData.view(_curr.buffer).getInt64(_pos, endian);
       _pos += 8;
       return result;
     }
   }
 
-  /// Reads and returns an 8-byte unsigned big-endian integer, converted to
-  /// a Dart int according to the semantics of ByteData.getUint64().
+  /// Reads and returns an 8-byte integer, converted to
+  /// a Dart int according to the semantics of [ByteData.getUint64]
+  /// using the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<int> readUnsignedLong() async {
     await _ensureNext();
-    if (_curr.length - _pos <= 8) {
+    if (_curr.length - _pos < 8) {
       // If we're on a buffer boundary
       // Keep it simple
-      return ByteData.view((await readBytes(8)).buffer).getUint64(0);
+      return ByteData.view((await readBytes(8)).buffer).getUint64(0, endian);
     } else {
-      var result = ByteData.view(_curr.buffer).getUint64(_pos);
+      var result = ByteData.view(_curr.buffer).getUint64(_pos, endian);
+      _pos += 8;
+      return result;
+    }
+  }
+
+  /// Reads and returns a 4-byte float, using the current [endian] setting.
+  ///
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
+  Future<double> readFloat() async {
+    await _ensureNext();
+    if (_curr.length - _pos < 4) {
+      // If we're on a buffer boundary
+      // Keep it simple
+      return ByteData.view((await readBytes(4)).buffer).getFloat32(0, endian);
+    } else {
+      var result = ByteData.view(_curr.buffer).getFloat32(_pos, endian);
+      _pos += 4;
+      return result;
+    }
+  }
+
+  /// Reads and returns an 8-byte double, using the current [endian] setting.
+  ///
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
+  Future<double> readDouble() async {
+    await _ensureNext();
+    if (_curr.length - _pos < 8) {
+      // If we're on a buffer boundary
+      // Keep it simple
+      return ByteData.view((await readBytes(8)).buffer).getFloat64(0, endian);
+    } else {
+      var result = ByteData.view(_curr.buffer).getFloat64(_pos, endian);
       _pos += 8;
       return result;
     }
   }
 
   /// Reads a string encoded in UTF8.  This method first reads a 2 byte
-  /// unsigned integer, given the number of UTF8 bytes to read.  It then reads
+  /// unsigned integer decoded with the current [endian] setting,
+  /// giving the number of UTF8 bytes to read.  It then reads
   /// those bytes, and converts them to a string using Dart's UTF8 conversion.
-  /// This may be incompatible with Java for a string that contains nulls.
+  /// This may be incompatible with `java.io.DataOutputStream`
+  /// for a string that contains nulls.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Future<String> readUTF8() async {
     final len = await readUnsignedShort();
     final bytes = await readBytesImmutable(len);
-    return utf8Decoder.convert(bytes);
+    return _utf8Decoder.convert(bytes);
   }
 
   /// Reads a byte, and returns false if it is 0, true otherwise.
   ///
-  /// Throws EOFException if EOF is reached before the needed byte is read.
+  /// Throws [EOFException] if EOF is reached before the needed byte is read.
   Future<bool> readBoolean() async {
     await _ensureNext();
     return _curr[_pos++] != 0;
@@ -336,21 +378,24 @@ class DataInputStream {
   }
 }
 
-/// ByteBufferDataInputStream is a wrapper around a ByteBuffer
+/// [ByteBufferDataInputStream] is a wrapper around a [ByteBuffer]
 /// instance.  This lets you synchronously read typed binary data from a byte
-/// array, much like java.io.DataInputStream over a ByteArrayInputStream.
+/// array, much like `java.io.DataInputStream` over a `ByteArrayInputStream`.
 ///
-/// See also DataInputStream
+/// See also [DataInputStream]
 class ByteBufferDataInputStream {
   final Uint8List _source;
   final ByteData _asByteData;
   int _pos = 0;
-  static const utf8Decoder = Utf8Decoder(allowMalformed: true);
+  static const _utf8Decoder = Utf8Decoder(allowMalformed: true);
 
-  ByteBufferDataInputStream._raw(this._source, this._asByteData);
+  /// The current endian setting, either [Endian.big] or [Endian.little].
+  /// This is used for converting numeric types.
+  Endian endian;
 
-  ByteBufferDataInputStream(Uint8List source)
-      : this._raw(source, source.buffer.asByteData());
+  ByteBufferDataInputStream(Uint8List source, [this.endian = Endian.big])
+      : _source = source,
+        _asByteData = source.buffer.asByteData();
 
   /// Returns the number of bytes that can be read from the internal buffer.
   ///  If we're at EOF, returns zero, otherwise, returns a non-zero positive
@@ -360,17 +405,17 @@ class ByteBufferDataInputStream {
   /// Check if we're at end of file.
   bool isEOF() => _source.lengthInBytes <= _pos;
 
-  /// Returns a new, mutable Uint8List containing the desired number
+  /// Returns a new, mutable [Uint8List] containing the desired number
   /// of bytes.
   Uint8List readBytes(int num) {
     return readBytesImmutable(num);
   }
 
-  /// Returns a potentially immutable Uint8List containing the desired number
+  /// Returns a potentially immutable [Uint8List] containing the desired number
   /// of bytes.  If the returned list is mutable, changing it might write
   /// through our internal buffer.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   Uint8List readBytesImmutable(int num) {
     if (_pos + num > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
@@ -381,53 +426,57 @@ class ByteBufferDataInputStream {
     }
   }
 
-  /// Reads and returns a 4-byte signed big-endian integer.
+  /// Reads and returns a 4-byte signed integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readInt() {
     if (_pos + 4 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getInt32(_pos);
+      final result = _asByteData.getInt32(_pos, endian);
       _pos += 4;
       return result;
     }
   }
 
-  /// Reads and returns a 4-byte unsigned big-endian integer.
+  /// Reads and returns a 4-byte unsigned integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readUnsignedInt() {
     if (_pos + 4 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getUint32(_pos);
+      final result = _asByteData.getUint32(_pos, endian);
       _pos += 4;
       return result;
     }
   }
 
-  /// Reads and returns a 2-byte unsigned big-endian integer.
+  /// Reads and returns a 2-byte unsigned integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readUnsignedShort() {
     if (_pos + 2 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getUint16(_pos);
+      final result = _asByteData.getUint16(_pos, endian);
       _pos += 2;
       return result;
     }
   }
 
-  /// Reads and returns a 2-byte signed big-endian integer.
+  /// Reads and returns a 2-byte signed integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readShort() {
     if (_pos + 2 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getInt16(_pos);
+      final result = _asByteData.getInt16(_pos, endian);
       _pos += 2;
       return result;
     }
@@ -435,8 +484,8 @@ class ByteBufferDataInputStream {
 
   /// Reads and returns an unsigned byte.
   ///
-  /// Throws EOFException if EOF is reached before the needed byte is read.
-  int readUnisgnedByte() {
+  /// Throws [EOFException] if EOF is reached before the needed byte is read.
+  int readUnsignedByte() {
     if (_pos + 1 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
@@ -446,63 +495,97 @@ class ByteBufferDataInputStream {
     }
   }
 
-  /// Reads and returns an unsigned byte.
+  /// Reads and returns a signed byte.  No sign extension is done.
   ///
-  /// Throws EOFException if EOF is reached before the needed byte is read.
+  /// Throws [EOFException] if EOF is reached before the needed byte is read.
   int readByte() {
     if (_pos + 1 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getInt8(_pos);
+      final b = _asByteData.getInt8(_pos);
       _pos += 1;
-      return result;
+      if (b & 0x80 != 0) {
+        return b - 0x100;
+      } else {
+        return b;
+      }
     }
   }
 
-  /// Reads and returns an 8-byte signed big-endian integer.
+  /// Reads and returns an 8-byte signed integer
+  /// decoded with the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readLong() {
     if (_pos + 8 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getInt64(_pos);
+      final result = _asByteData.getInt64(_pos, endian);
       _pos += 8;
       return result;
     }
   }
 
-  /// Reads and returns an 8-byte unsigned big-endian integer, converted to
-  /// a Dart int according to the semantics of ByteData.getUint64().
+  /// Reads and returns an 8-byte unsigned integer, converted to
+  /// a Dart int according to the semantics of [ByteData.getUint64]
+  /// using the current [endian] setting.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   int readUnsignedLong() {
     if (_pos + 8 > _source.lengthInBytes) {
       throw EOFException('Attempt to read beyond end of input');
     } else {
-      final result = _asByteData.getUint64(_pos);
+      final result = _asByteData.getUint64(_pos, endian);
+      _pos += 8;
+      return result;
+    }
+  }
+
+  /// Reads and returns a 4-byte float, using the current [endian] setting.
+  ///
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
+  double readFloat() {
+    if (_pos + 4 > _source.lengthInBytes) {
+      throw EOFException('Attempt to read beyond end of input');
+    } else {
+      final result = _asByteData.getFloat32(_pos, endian);
+      _pos += 4;
+      return result;
+    }
+  }
+
+  /// Reads and returns an 8-byte double, using the current [endian] setting.
+  ///
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
+  double readDouble() {
+    if (_pos + 8 > _source.lengthInBytes) {
+      throw EOFException('Attempt to read beyond end of input');
+    } else {
+      final result = _asByteData.getFloat64(_pos, endian);
       _pos += 8;
       return result;
     }
   }
 
   /// Reads a string encoded in UTF8.  This method first reads a 2 byte
-  /// unsigned integer, given the number of UTF8 bytes to read.  It then reads
+  /// unsigned integer decoded with the current [endian] setting,
+  /// giving the number of UTF8 bytes to read.  It then reads
   /// those bytes, and converts them to a string using Dart's UTF8 conversion.
-  /// This may be incompatible with Java for a string that contains nulls.
+  /// This may be incompatible with `java.io.DataOutputStream` for a string
+  /// that contains nulls.
   ///
-  /// Throws EOFException if EOF is reached before the needed bytes are read.
+  /// Throws [EOFException] if EOF is reached before the needed bytes are read.
   String readUTF8() {
     final len = readUnsignedShort();
     final bytes = readBytesImmutable(len);
-    return utf8Decoder.convert(bytes);
+    return _utf8Decoder.convert(bytes);
   }
 
-  /// Reads a byte, and returns false if it is 0, true otherwise.
+  /// Reads a byte, and returns `false` if it is 0, `true` otherwise.
   ///
   /// Throws EOFException if EOF is reached before the needed byte is read.
   bool readBoolean() {
-    return readByte() != 0;
+    return readUnsignedByte() != 0;
   }
 
   /// Returns the next unsigned byte, or -1 on EOF
@@ -510,7 +593,7 @@ class ByteBufferDataInputStream {
     if (isEOF()) {
       return -1;
     } else {
-      return readByte();
+      return readUnsignedByte();
     }
   }
 
@@ -520,22 +603,26 @@ class ByteBufferDataInputStream {
   }
 }
 
-/// An adapter over a Sink<List<int>> for writing to a stream.  This class
-/// presents an API based on Java's java.io.DataOutputStream.
+/// An adapter over a `Sink<List<int>>` for writing to a stream.  This class
+/// presents an API based on Java's `java.io.DataOutputStream`.
 ///
-/// DataOutputStream does no buffering.
+/// DataOutputSink does no buffering.
 class DataOutputSink {
   final Sink<List<int>> _dest;
 
-  DataOutputSink(this._dest);
+  /// The current endian setting, either [Endian.big] or [Endian.little].
+  /// This is used for converting numeric types.
+  Endian endian;
+
+  DataOutputSink(this._dest, [this.endian = Endian.big]);
 
   /// Write the low 8 bits of the given integer as a single byte.
   void writeByte(int b) {
     _dest.add(Uint8List.fromList([b & 0xff]));
   }
 
-  /// Write the given bytes.  A Uint8List is the recommended subtype
-  /// of List<int>.
+  /// Write the given bytes.  A [Uint8List] is the recommended subtype
+  /// of `List<int>`.
   void writeBytes(List<int> bytes) {
     if (bytes is Uint8List) {
       _dest.add(bytes);
@@ -544,51 +631,66 @@ class DataOutputSink {
     }
   }
 
-  /// Writes a byte containg 0 if v is false, and 1 if it is true.
+  /// Writes a byte containg `0` if `v` is `false`, and `1` if it is `true`.
   void writeBoolean(bool v) {
     writeByte(v ? 1 : 0);
   }
 
-  /// Writes a 2 byte signed short in big-endian format.
+  /// Writes a 2 byte signed short using the current [endian] format.
   void writeShort(int v) {
-    final d = ByteData(2)..setInt16(0, v);
+    final d = ByteData(2)..setInt16(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Writes a 2 byte unsigned short in big-endian format.
+  /// Writes a 2 byte unsigned short using the current [endian] format.
   void writeUnsignedShort(int v) {
-    final d = ByteData(2)..setUint16(0, v);
+    final d = ByteData(2)..setUint16(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Writes a 4 byte int in big-endian format.
+  /// Writes a 4 byte int using the current [endian] format.
   void writeInt(int v) {
-    final d = ByteData(4)..setInt32(0, v);
+    final d = ByteData(4)..setInt32(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Writes a 4 byte unsigned int in big-endian format.
+  /// Writes a 4 byte unsigned int using the current [endian] format.
   void writeUnsignedInt(int v) {
-    final d = ByteData(4)..setUint32(0, v);
+    final d = ByteData(4)..setUint32(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Writes an 8 byte long int in big-endian format.
+  /// Writes an 8 byte long int using the current [endian] format.
   void writeLong(int v) {
-    final d = ByteData(8)..setInt64(0, v);
+    final d = ByteData(8)..setInt64(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Writes an unsigned 8 byte long in big-endian format.  Converts a Dart
-  /// int according to the semantics of ByteData.setUint64().
+  /// Writes an unsigned 8 byte long using the current [endian] format.
+  /// Converts a Dart
+  /// `int` according to the semantics of [ByteData.setUint64`.
   void writeUnsignedLong(int v) {
-    final d = ByteData(8)..setUint64(0, v);
+    final d = ByteData(8)..setUint64(0, v, endian);
     _dest.add(d.buffer.asUint8List());
   }
 
-  /// Write the given string as an unsigned short byte length, followed
-  /// by an array of UTF8 characters.  This shoould be compatable with
-  /// Java's DataOutputStream.writeUTF8, as long as the string contains no
+  /// Writes a four-byte float using the current [endian] format.
+  void writeFloat(double v) {
+    final d = ByteData(4)..setFloat32(0, v, endian);
+    _dest.add(d.buffer.asUint8List());
+  }
+
+  /// Writes an eight-byte double using the current [endian] format.
+  void writeDouble(double v) {
+    final d = ByteData(8)..setFloat64(0, v, endian);
+    _dest.add(d.buffer.asUint8List());
+  }
+
+  /// Write the given string as an unsigned short giving the byte length,
+  /// followed by that many bytes containing the UTF8 encoding.
+  /// The short is written using the current [endian] format.
+  /// This should be compatable with
+  /// Java's `DataOutputStream.writeUTF8()`, as long as the string contains no
   /// nulls and no UTF formats beyond the 1, 2 and 3 byte formats.  See
   /// https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/io/DataInput.html#modified-utf-8
   ///
@@ -609,8 +711,8 @@ class DataOutputSink {
   }
 }
 
-/// A wrapper around a DataInputStream that decrypts using
-/// a Pointycastle BlockCipher, resulting in a Stream<List<int>>.
+/// A wrapper around a [DataInputStream] that decrypts using
+/// a Pointycastle [BlockCipher], resulting in a `Stream<List<int>>`.
 class DecryptingStream extends DelegatingStream<Uint8List> {
   DecryptingStream(BlockCipher cipher, Stream<List<int>> input, Padding padding)
       : this.fromDataInputStream(cipher, DataInputStream(input), padding);
@@ -647,9 +749,9 @@ class DecryptingStream extends DelegatingStream<Uint8List> {
   }
 }
 
-/// A wrapper around a Sink<List<int>> that encrypts data using
-/// a Pointycastle BlockCipher.  This sink needs to buffer to build
-/// up blocks that can be encrypted, so it's essential that close()
+/// A wrapper around a `Sink<List<int>>` that encrypts data using
+/// a Pointycastle [BlockCipher].  This sink needs to buffer to build
+/// up blocks that can be encrypted, so it's essential that [close()]
 /// be called.
 class EncryptingSink implements Sink<List<int>> {
   final BlockCipher _cipher;