+++ dart/jovial_misc/CHANGELOG.md
@@ -13,3 +13,14 @@
 ## 0.1.1
 
 - Cosmetic fixes to README and example
+
+## 0.2.0
+
+- Added the ability to change the endianness of DataInputStream
+  and DataOutputSink.
+- Added float and double to DataInputStream and DataOutputSink.
+- Fixed ByteBufferDataInputStream.readByte() to return a signed
+  byte, as intended.
+- Made the unintentionally public DataInputStream.utf8Decoder
+  into a private value.
+- Documentation improvements.